+++ CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [5.5.0] - 2026-06-29
+
+### Fixed
+
+- Disable inline event handlers on the login/logout forms to comply with the CSP from `jeap-spring-boot-web-config-starter`
+- Disable browser caching of the login/logout forms by overriding the starter's `Cache-Control` headers
+
 ## [5.4.0] - 2026-06-23
 
 ### Changed